requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
supabase
google-generativeai>=0.8.0
pymupdf==1.23.8
python-dotenv==1.0.0
pydantic==2.6.0
pydantic-settings==2.1.0
requests==2.31.0
pytest==8.0.0
pytest-asyncio==0.23.5
httpx>=0.27.0
supabase>=2.12.0

gunicorn==21.2.0
